requirements.txt
langgraph
google-generativeai
pydantic
streamlit